First Jest test with Enzyme -- testing PhotoGallery Component rendering
