+++ package.json
@@ -11,7 +11,19 @@
     "start": "nodemon ./server/index.js",
     "dev": "webpack -d --watch",
     "build": "webpack -p",
-    "seed": "node ./db/seed-db.js"
+    "seed": "node ./db/seed-db.js",
+    "test": "jest",
+    "test:watch": "jest --watch",
+    "test:coverage": "jest --coverage"
+  },
+  "jest": {
+    "setupFiles": [
+      "./client/src/jest-setup.js"
+    ],
+    "snapshotSerializers": [
+      "enzyme-to-json/serializer"
+    ],
+    "verbose": true
   },
   "babel": {
     "presets": [
@@ -36,6 +48,9 @@
     "babel-preset-env": "^1.6.1",
     "babel-preset-react": "^6.24.1",
     "babel-preset-stage-2": "^6.24.1",
+    "enzyme": "^3.3.0",
+    "enzyme-adapter-react-16": "^1.1.1",
+    "enzyme-to-json": "^3.3.1",
     "eslint": "^4.18.1",
     "eslint-config-airbnb": "^16.1.0",
     "eslint-config-airbnb-base": "^12.1.0",
@@ -44,6 +59,7 @@
     "eslint-plugin-react": "^7.7.0",
     "jest": "^22.4.0",
     "puppeteer": "^1.1.0",
+    "react-test-renderer": "^16.2.0",
     "readline": "^1.3.0",
     "readline-async": "^0.1.0",
     "webpack": "^3.11.0",
